refactor(workflows): split WorkflowsController into per-concern controllers, dedup ownership guard + error handling (#836)
workflows.controller.ts was a single 1,467-line controller with 40 routes
across 7+ concerns, a 13x-duplicated ownership-guard preamble (with three
divergent 404 shapes), and a 14x-duplicated try/catch->500 idiom (most copies
silently downgrading service-thrown 400/404s to 500).

Split along the section banners into focused controllers, all under the same
`workflows` route prefix (paths unchanged):
- WorkflowBuilderController  — nodes registry, validate-connection, generate,
  import, interface, validate(-reference)
- WorkflowMarketplaceController — templates, marketplace, referencable,
  publish/unpublish
- WorkflowExecutionController — schedule, execute/resume, credits-estimate,
  lifecycle, execution logs, review-gate approval, node lock/unlock, thumbnail
- WorkflowBatchController — batch status/list/run
- WorkflowWebhookManagementController — per-workflow webhook config
- WorkflowCrudController — create/list/statistics/find/clone/update/remove +
  export-comfyui

Dedup + correctness:
- WorkflowsService.findOwnedOrThrow(workflowId, { organization, user? }) is the
  single ownership guard; every inline preamble now delegates to it and 404s
  use one canonical NotFoundException('Workflow not found').
- New wrapError() util replaces the 14 try/catch->500 copies; it always
  rethrows existing HttpExceptions, so a service-thrown 400/404 (e.g. from
  import / executePartial / submitApproval) reaches the client with its real
  status instead of a generic 500. Covered by tests.
- getWebhookInfo no longer reads process.env.API_URL directly; added a
  ConfigService.apiUrl getter (GENFEEDAI_API_URL) and use it.
- getBatchStatus/listBatchJobs inline response types moved to
  packages/interfaces (BatchWorkflowStatusResponse / BatchWorkflowSummary).

Module registers each new controller, ordering literal-route controllers
before WorkflowCrudController so `:workflowId` never shadows templates /
statistics / batch / marketplace routes. All 40 routes + verbs unchanged.

Closes #836
