+++ packages/interfaces/src/batch/index.ts
@@ -1,2 +1,3 @@
 export * from './batch.interface';
+export * from './batch-workflow-status.interface';
 export * from './manual-review-batch-item.interface';
